+++ CHANGELOG.md
@@ -5,9 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## v2.1.0 - 2025-09-23
 ### What's Changed
-**Full Changelog**: https://github.com/obervinov/terraform-setup-environment/compare/v2.0.0...v2.1.0 by @obervinov
+**Full Changelog**: https://github.com/obervinov/terraform-setup-environment/compare/v2.0.1...v2.1.0 by @obervinov
 #### 🚀 Features
 * added an additional parameter `droplet_image_id` to support direct image ID usage instead of resolving the image slug (useful for cases with imported legacy droplets that reference images no longer available in the DO catalog. Parameter used as mocked in the `data.digitalocean_droplet_snapshot` data source to avoid resolution errors)
+* added an additional parameter `droplet_name_override` to support custom droplet name formats (useful for cases with imported legacy droplets that do not follow the default naming convention with region suffix). This is a temporary parameter for compatibility with existing setups and will be removed in future releases.
+* added an additional parameter `droplet_provisioner_ssh_key_name` to specify the name of the SSH key in DigitalOcean for provisioner connection to droplet to execute remote-exec (default: "terraform")
 #### 💥 Breaking Changes
 * default value of the `droplet_image` variable changed from `ubuntu-24-04.rev1` to `ubuntu-22-04-x64` from public catalog (to avoid confusion with private snapshots that may not be available for all users)